+++ src/process.h
@@ -13,14 +13,14 @@
 class SnapshotProcessor{
 private:
 public:
-  std::vector<float> a1 = std::vector<float>(3); // Lattice vectors
-  std::vector<float> a2 = std::vector<float>(3);
-  std::vector<float> a3 = std::vector<float>(3);
-  std::vector<std::vector<float>> positions; // Positions of particles in box
-  std::vector<std::vector<float>> q;
-  std::vector<std::vector<float>> w;
-  std::vector<std::vector<float>> q_av;
-  std::vector<std::vector<float>> w_av;
+  std::vector<double> a1 = std::vector<double>(3); // Lattice vectors
+  std::vector<double> a2 = std::vector<double>(3);
+  std::vector<double> a3 = std::vector<double>(3);
+  std::vector<std::vector<double>> positions; // Positions of particles in box
+  std::vector<std::vector<double>> q;
+  std::vector<std::vector<double>> w;
+  std::vector<std::vector<double>> q_av;
+  std::vector<std::vector<double>> w_av;
   SnapshotProcessor(void){};
   virtual ~SnapshotProcessor(void){};
 
@@ -33,7 +33,7 @@ public:
   void save_boops(
     const std::string target_dir,
     const std::string file_name,
-    std::vector<std::vector<float>>& boop_vector
+    std::vector<std::vector<double>>& boop_vector
   )const;
 };
